Factored out common code accross different controllers in controller.go
