--- pkg/controllers/controller/controller.go
@@ -1,7 +1,0 @@
-package controller
-
-// Controller interface
-type Controller interface {
-	// Run method
-	Run(threadiness int, reconcilerPeriod string, stopCh chan struct{})
-}
