+++ pkg/controllers/namespace_controller.go
@@ -1,15 +1,13 @@
-package namespace
+package controllers
 
 import (
 	calicocache "github.com/projectcalico/k8s-policy/pkg/cache"
-	"github.com/projectcalico/k8s-policy/pkg/controllers/controller"
 	"github.com/projectcalico/k8s-policy/pkg/converter"
 	"github.com/projectcalico/libcalico-go/lib/api"
 	"github.com/projectcalico/libcalico-go/lib/client"
 	"github.com/projectcalico/libcalico-go/lib/errors"
 	log "github.com/sirupsen/logrus"
 	"k8s.io/apimachinery/pkg/fields"
-	uruntime "k8s.io/apimachinery/pkg/util/runtime"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/pkg/api/v1"
 	"k8s.io/client-go/tools/cache"
@@ -21,15 +19,11 @@ import (
 // Responsible for monitoring kubernetes namespaces and
 // syncing them to Calico datastore.
 type NamespaceController struct {
-	indexer        cache.Indexer
-	informer       cache.Controller
-	calicoObjCache calicocache.ResourceCache
-	calicoClient   *client.Client
-	k8sClientset   *kubernetes.Clientset
+	BaseController
 }
 
 // NewNamespaceController Constructor for NamespaceController
-func NewNamespaceController(k8sClientset *kubernetes.Clientset, calicoClient *client.Client) controller.Controller {
+func NewNamespaceController(k8sClientset *kubernetes.Clientset, calicoClient *client.Client) Controller {
 
 	namespaceConverter := converter.NewNamespaceConverter()
 
@@ -144,123 +138,49 @@ func NewNamespaceController(k8sClientset *kubernetes.Clientset, calicoClient *cl
 		},
 	}, cache.Indexers{})
 
-	return &NamespaceController{indexer, informer, ccache, calicoClient, k8sClientset}
-}
+	// function syncs the given update to Calico's datastore,
+	datastoreWriteFunc := func(key string) error {
+		// Check if it exists in our cache.
+		obj, exists := ccache.Get(key)
 
-// Run starts controller.Internally it starts syncing
-// kubernetes and calico caches.
-func (c *NamespaceController) Run(threadiness int, reconcilerPeriod string, stopCh chan struct{}) {
-	defer uruntime.HandleCrash()
+		if !exists {
 
-	// Let the workers stop when we are done
-	workqueue := c.calicoObjCache.GetQueue()
-	defer workqueue.ShutDown()
+			log.Debugf("namespace %s does not exist anymore on kubernetes\n", key)
+			log.Infof("Deleting namespace %s on ETCD \n", key)
 
-	log.Info("Starting namespace controller")
+			err := calicoClient.Profiles().Delete(api.ProfileMetadata{
+				Name: key,
+			})
 
-	// Start Calico cache. Cache gets loaded with objects
-	// from ETCD datastore.
-	c.calicoObjCache.Run(reconcilerPeriod)
+			// Let Delete() operation be idompotent. Ignore the error while deletion if
+			// object does not exists on ETCD already.
+			if err != nil {
+				if _, ok := err.(errors.ErrorResourceDoesNotExist); !ok {
+					log.WithError(err).Errorf("Got error while deleting %s in datastore.", key)
+					return err
+				}
+			}
+			return nil
+		} else {
 
-	go c.informer.Run(stopCh)
+			var p api.Profile
+			p = obj.(api.Profile)
+			log.Infof("Applying namespace %s on ETCD \n", key)
+			_, err := calicoClient.Profiles().Apply(&p)
 
-	// Wait till k8s cache is synced
-	for !c.informer.HasSynced() {
-	}
-
-	// Start a number of worker threads to read from the queue.
-	for i := 0; i < threadiness; i++ {
-		go c.runWorker()
-	}
-
-	<-stopCh
-	log.Info("Stopping Node controller")
-}
-
-func (c *NamespaceController) runWorker() {
-	for c.processNextItem() {
-	}
-}
-
-func (c *NamespaceController) processNextItem() bool {
-	// Wait until there is a new item in the working queue
-	workqueue := c.calicoObjCache.GetQueue()
-	key, quit := workqueue.Get()
-	if quit {
-		return false
-	}
-
-	// Tell the queue that we are done with processing this key. This unblocks the key for other workers
-	// This allows safe parallel processing because two nodes with the same key are never processed in
-	// parallel.
-	defer workqueue.Done(key)
-
-	// Invoke the method containing the business logic
-	err := c.syncToCalico(key.(string))
-
-	// Handle the error if something went wrong during the execution of the business logic
-	c.handleErr(err, key.(string))
-	return true
-}
-
-// syncToCalico syncs the given update to Calico's etcd, as well as the in-memory cache
-// of Calico objects.
-func (c *NamespaceController) syncToCalico(key string) error {
-	// Check if it exists in our cache.
-	obj, exists := c.calicoObjCache.Get(key)
-
-	if !exists {
-
-		log.Debugf("namespace %s does not exist anymore on kubernetes\n", key)
-		log.Infof("Deleting namespace %s on ETCD \n", key)
-
-		err := c.calicoClient.Profiles().Delete(api.ProfileMetadata{
-			Name: key,
-		})
-
-		// Let Delete() operation be idompotent. Ignore the error while deletion if
-		// object does not exists on ETCD already.
-		if _, ok := err.(errors.ErrorResourceDoesNotExist); ok {
-			err = nil
+			return err
 		}
-
-		return err
-	} else {
-
-		var p api.Profile
-		p = obj.(api.Profile)
-		log.Infof("Applying namespace %s on ETCD \n", key)
-		_, err := c.calicoClient.Profiles().Apply(&p)
-
-		return err
-	}
-}
-
-// handleErr checks if an error happened and makes sure we will retry later.
-func (c *NamespaceController) handleErr(err error, key string) {
-	workqueue := c.calicoObjCache.GetQueue()
-	if err == nil {
-
-		// Forget about the #AddRateLimited history of the key on every successful synchronization.
-		// This ensures that future processing of updates for this key is not delayed because of
-		// an outdated error history.
-		workqueue.Forget(key)
-		return
 	}
 
-	// This controller retries 5 times if something goes wrong. After that, it stops trying.
-	if workqueue.NumRequeues(key) < 5 {
-
-		log.WithError(err).Errorf("Error syncing namespace %v: %v", key, err)
-		// Re-enqueue the key rate limited. Based on the rate limiter on the
-		// queue and the re-enqueue history, the key will be processed later again.
-		workqueue.AddRateLimited(key)
-		return
+	return &NamespaceController{
+		BaseController{
+			indexer:            indexer,
+			informer:           informer,
+			calicoObjCache:     ccache,
+			calicoClient:       calicoClient,
+			k8sClientset:       k8sClientset,
+			controllerType:     "NamespaceController",
+			datastoreWriteFunc: datastoreWriteFunc,
+		},
 	}
-
-	workqueue.Forget(key)
-
-	// Report to an external entity that, even after several retries, we could not successfully process this key
-	uruntime.HandleError(err)
-	log.WithError(err).Errorf("Dropping namespace %q out of the queue: %v", key, err)
 }
