+++ pkg/controllers/pod_controller.go
@@ -1,8 +1,7 @@
-package pod
+package controllers
 
 import (
 	calicocache "github.com/projectcalico/k8s-policy/pkg/cache"
-	"github.com/projectcalico/k8s-policy/pkg/controllers/controller"
 	"github.com/projectcalico/k8s-policy/pkg/converter"
 	"github.com/projectcalico/libcalico-go/lib/api"
 	"github.com/projectcalico/libcalico-go/lib/client"
@@ -20,16 +19,12 @@ import (
 // Responsible for monitoring kubernetes pods and
 // updating pod labels to calico ETCD if changed.
 type PodController struct {
-	indexer       cache.Indexer
-	informer      cache.Controller
-	labelCache    calicocache.ResourceCache
+	BaseController
 	endpointCache map[string]api.WorkloadEndpoint
-	calicoClient  *client.Client
-	k8sClientset  *kubernetes.Clientset
 }
 
 // NewPodController Constructor for PodController
-func NewPodController(k8sClientset *kubernetes.Clientset, calicoClient *client.Client) controller.Controller {
+func NewPodController(k8sClientset *kubernetes.Clientset, calicoClient *client.Client) Controller {
 	podConverter := converter.NewPodConverter()
 
 	// Function returns map of key->workloadEndpoint stored by policy controller
@@ -62,7 +57,7 @@ func NewPodController(k8sClientset *kubernetes.Clientset, calicoClient *client.C
 	}
 
 	labelCache := calicocache.NewResourceCache(cacheArgs)
-	endpointCache := make(map[string]api.WorkloadEndpoint)
+	endpointCache := map[string]api.WorkloadEndpoint{}
 
 	// create the watcher
 	listWatcher := cache.NewListWatchFromClient(k8sClientset.Core().RESTClient(), "pods", "", fields.Everything())
@@ -139,7 +134,81 @@ func NewPodController(k8sClientset *kubernetes.Clientset, calicoClient *client.C
 		},
 	}, cache.Indexers{})
 
-	return &PodController{indexer, informer, labelCache, endpointCache, calicoClient, k8sClientset}
+	// function syncs the given update to Calico's datastore,
+	datastoreWriteFunc := func(key string) error {
+		// Check if it exists in our cache.
+		if labels, exists := labelCache.Get(key); exists {
+
+			// Get workloadEndpoint from cache
+			endpoint, exists := endpointCache[key]
+
+			if !exists {
+				log.Errorf("no endpoint for %s in cache, loading", key)
+
+				// Load endpoint cache. Retry when failed.
+				err := populateEndpointCache()
+				if err != nil {
+					log.WithError(err).Error("failed to load workload endpoint cache")
+					return err
+				}
+
+				endpoint, exists = endpointCache[key]
+				if !exists {
+
+					// No endpoint in etcd - this means the pod hasn't been
+					// created by the CNI plugin yet. Just wait until it has been.
+					// This can only be hit when labels for a pod change before
+					//  the pod has been deployed, so should be pretty uncommon.
+					log.Infof("Pod %s hasn't been created by the CNI plugin yet.", key)
+					return nil
+				}
+			}
+
+			oldLabels := endpoint.Metadata.Labels
+			newLabels := labels.(map[string]string)
+
+			if !reflect.DeepEqual(oldLabels, newLabels) {
+
+				// Update the labels on endpoint
+				endpoint.Metadata.Labels = newLabels
+
+				log.Infof("Updating endpoint %s with labels %#v on ETCD \n", key, newLabels)
+				_, err := calicoClient.WorkloadEndpoints().Update(&endpoint)
+
+				if err != nil {
+					log.WithError(err).Errorf("failed to update workload endpoint %s", key)
+					return err
+				} else {
+
+					// Update endpoint cache as well with modified endpoint.
+					endpointCache[key] = endpoint
+				}
+
+				if _, ok := err.(errors.ErrorResourceDoesNotExist); ok {
+
+					// Endpoint not yet created by CNI plugin.
+					err = nil
+				}
+
+				return err
+			}
+		}
+
+		return nil
+	}
+
+	return &PodController{
+		BaseController: BaseController{
+			indexer:            indexer,
+			informer:           informer,
+			calicoObjCache:     labelCache,
+			calicoClient:       calicoClient,
+			k8sClientset:       k8sClientset,
+			controllerType:     "PodController",
+			datastoreWriteFunc: datastoreWriteFunc,
+		},
+		endpointCache: endpointCache,
+	}
 }
 
 // Run starts controller.Internally it starts syncing
@@ -148,14 +217,14 @@ func (c *PodController) Run(threadiness int, reconcilerPeriod string, stopCh cha
 	defer uruntime.HandleCrash()
 
 	// Let the workers stop when we are done
-	workqueue := c.labelCache.GetQueue()
+	workqueue := c.BaseController.calicoObjCache.GetQueue()
 	defer workqueue.ShutDown()
 
 	log.Info("Starting pod controller")
 
 	// Start Calico cache. Cache gets loaded with objects
-	// from ETCD datastore.
-	c.labelCache.Run(reconcilerPeriod)
+	// from Calico datastore.
+	c.BaseController.calicoObjCache.Run(reconcilerPeriod)
 
 	// Load endpoint cache. Retry when failed.
 	for err := c.populateEndpointCache(); err != nil; {
@@ -175,123 +244,6 @@ func (c *PodController) Run(threadiness int, reconcilerPeriod string, stopCh cha
 
 	<-stopCh
 	log.Info("Stopping Pod controller")
-}
-
-func (c *PodController) runWorker() {
-	for c.processNextItem() {
-	}
-}
-
-func (c *PodController) processNextItem() bool {
-	// Wait until there is a new item in the working queue
-	workqueue := c.labelCache.GetQueue()
-	key, quit := workqueue.Get()
-	if quit {
-		return false
-	}
-
-	// Tell the queue that we are done with processing this key. This unblocks the key for other workers
-	// This allows safe parallel processing because two nodes with the same key are never processed in
-	// parallel.
-	defer workqueue.Done(key)
-
-	// Invoke the method containing the business logic
-	err := c.syncToCalico(key.(string))
-
-	// Handle the error if something went wrong during the execution of the business logic
-	c.handleErr(err, key.(string))
-	return true
-}
-
-// syncToCalico syncs the given update to Calico's etcd
-func (c *PodController) syncToCalico(key string) error {
-	// Check if it exists in our cache.
-	if labels, exists := c.labelCache.Get(key); exists {
-
-		// Get workloadEndpoint from cache
-		endpoint, exists := c.endpointCache[key]
-
-		if !exists {
-			log.Errorf("no endpoint for %s in cache, loading", key)
-
-			// Load endpoint cache. Retry when failed.
-			err := c.populateEndpointCache()
-			if err != nil {
-				log.WithError(err).Error("failed to load workload endpoint cache")
-				return err
-			}
-
-			endpoint, exists = c.endpointCache[key]
-			if !exists {
-
-				// No endpoint in etcd - this means the pod hasn't been
-				// created by the CNI plugin yet. Just wait until it has been.
-				// This can only be hit when labels for a pod change before
-				//  the pod has been deployed, so should be pretty uncommon.
-				log.Infof("Pod %s hasn't been created by the CNI plugin yet.", key)
-				return nil
-			}
-		}
-
-		oldLabels := endpoint.Metadata.Labels
-		newLabels := labels.(map[string]string)
-
-		if !reflect.DeepEqual(oldLabels, newLabels) {
-
-			// Update the labels on endpoint
-			endpoint.Metadata.Labels = newLabels
-
-			log.Infof("Updating endpoint %s with labels %#v on ETCD \n", key, newLabels)
-			_, err := c.calicoClient.WorkloadEndpoints().Update(&endpoint)
-
-			if err != nil {
-				log.WithError(err).Errorf("failed to update workload endpoint %s", key)
-				return err
-			}else{
-
-				// Update endpoint cache as well with modified endpoint.
-				c.endpointCache[key] = endpoint
-			}
-
-			if _, ok := err.(errors.ErrorResourceDoesNotExist); ok {
-
-				// Endpoint not yet created by CNI plugin.
-				err = nil
-			}
-
-			return err
-		}
-	}
-
-	return nil
-}
-
-// handleErr checks if an error happened and makes sure we will retry later.
-func (c *PodController) handleErr(err error, key string) {
-	workqueue := c.labelCache.GetQueue()
-	if err == nil {
-
-		// Forget about the #AddRateLimited history of the key on every successful synchronization.
-		// This ensures that future processing of updates for this key is not delayed because of
-		// an outdated error history.
-		workqueue.Forget(key)
-		return
-	}
-
-	// This controller retries 5 times if something goes wrong. After that, it stops trying.
-	if workqueue.NumRequeues(key) < 5 {
-		log.WithError(err).Errorf("Error syncing pod %v: %v", key, err)
-		// Re-enqueue the key rate limited. Based on the rate limiter on the
-		// queue and the re-enqueue history, the key will be processed later again.
-		workqueue.AddRateLimited(key)
-		return
-	}
-
-	workqueue.Forget(key)
-	
-	// Report to an external entity that, even after several retries, we could not successfully process this key
-	uruntime.HandleError(err)
-	log.WithError(err).Errorf("Dropping pod %q out of the queue: %v", key, err)
 }
 
 // populateEndpointCache() Loads map of workload endpoint objects from ETCD datastore
